STT2024/requirements.txt
pyaudio
pybind11
openai-whisper
vosk
pyttsx3
paho-mqtt
configobj
snakeviz
